Support higher get routing table command interval
For some of the validation tests, it is quite useful to adjust the
successive getRoutingTable command interval to a high value.
Current size of uint8_t provides a interval of 255s whereas uint16_t
provides a interval of 18 hours.

Tested:
Attempted a higher interval using modified configurations, worked as
per expectation.

+++ include/utils/Configuration.hpp
@@ -55,7 +55,7 @@ struct I3CConfiguration : Configuration
     uint8_t I3CAddress = 0;
     bool requiresCpuPidMask = true;
     uint8_t provisionalIdMask;
-    uint8_t getRoutingInterval = 0;
+    uint16_t getRoutingInterval = 0;
     bool forwaredEIDPoolToEP = false;
     bool blockDiscoveryNotify = false;
